+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - Split build into buildNodes and buildScopes
 - Allow to override nodes and scopes while instantiating a scope blue print
 - Define overrides in constructor
+- Refactor ScopePlugin
 
 ### Fixed